backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
redis-om==0.3.2
redis==5.2.0
openai==1.54.0
python-dotenv==1.0.1
pydantic==2.9.2
websockets==13.1
